requirements.txt
charade==1.0.3
JPype1-py3==0.5.5.2
numpy==1.14.2
scikit-learn==0.19.1
scipy==1.0.1
